tests/wpt/tests/speculation-rules/prerender/cancel-prerendering-after-clear-site-data-cache.https.html

<!DOCTYPE html>
<title>clear-site-data-cache cancels prerenders</title>
<meta name="timeout" content="long">
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/speculation-rules/resources/utils.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/html/browsers/browsing-the-web/remote-context-helper/resources/remote-context-helper.js"></script>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {

  const rcHelper = new RemoteContextHelper();
  const referrerRC = await rcHelper.addWindow({
    origin: 'HTTPS_ORIGIN'
  }, {
    features: 'noopener'
  });
  const prerenderedRC = await addPrerenderRC(referrerRC, {
    origin: 'HTTPS_REMOTE_ORIGIN',
    headers: [
      ['Supports-Loading-Mode', 'credentialed-prerender']
    ]
  });

  // Open new window with url to clear cache data through Clear-Site-Data header.
  // Ensure that the cache is cleared before the navigation.
  const gotMessage = new Promise(resolve => {
    window.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  let cache_helper = "cache_helper=" + self.crypto.randomUUID() + "&";
  new_window = window.open("../../clear-site-data/support/clear-site-data-cache.py?" + cache_helper + "response=single_html&cache&clear_first=all");
  await gotMessage;

  // Because Clear-Site-Header response header is sent, the existing prerender
  // is expected to be canceled.
  // And the navigation is expected to create another page instead of activation.
  referrerRC.navigateTo(prerenderedRC.url);
  assert_equals(await getActivationStart(prerenderedRC), 0);
});
</script>
</body>
</html>
